+++ examples/async/s3_cp.ml
@@ -403,4 +403,4 @@ let () =
       +> anon ("src" %: string)
       +> anon ("dst" %: string))
     main
-  |> run
+  |> Command_unix.run