+++ CHANGES.md
@@ -32,6 +32,7 @@
 - New lwt based server implementation: cohttp-server-lwt-unix
   + This new implementation does not depend on conduit and has a simpler and
     more flexible API
+- async: Adapt cohttp-curl-async to work with core_unix.
 - *Breaking changes*
   + refactor: move opam metadata to dune-project (rgrinberg #811)
   + refactor: deprecate Cohttp_async.Io (rgrinberg #807)